Block delete reservation after its start time
If start time of equipment or place passed current time, then block delete function

+++ src/popo/reservation/equip/reserve.equip.controller.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   Body,
   Controller,
   Delete,
@@ -169,6 +170,9 @@ export class ReserveEquipController {
       await this.reserveEquipService.remove(uuid);
     } else {
       if (reservation.booker_id == user.uuid) {
+        if(reservation.start_time < new Date().toISOString()) {
+          throw new BadRequestException('Cannot delete past reservation');
+        }
         await this.reserveEquipService.remove(uuid);
       } else {
         throw new UnauthorizedException('Unauthorized delete action');
